pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
numpy = "^1.26.2"
lightning = "^2.1.3"
torch = "^2.1.2"
torchvision = "^0.16.2"
hydra-core = "^1.3.2"
pandas = "^2.1.4"
fire = "^0.5.0"
gitpython = "^3.1.40"
mlflow = "^2.9.2"
dvc = "^3.37.0"
dvc-gdrive = "^3.0.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
tritonclient = {extras = ["all"], version = "^2.41.0"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py37"]
skip-string-normalization = true


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
